third_party/blink/web_tests/wpt_internal/css/css-conditional/container-queries/container-selection.html

<!DOCTYPE html>
<title>@container: selection using name(), type()</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-rule">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>

  main { background-color: lightgray; }
  main > div { background-color: skyblue; }
  main > div > div { background-color: seagreen; }
  main > div > div > div { background-color: tomato; }

  main {
    width: 64px;
    height: 64px;
  }

  main div {
    width: 50%;
    height: 50%;
  }

  .inline { container-type: inline-size; }
  .block { container-type: block-size; }
  .size { container-type: size; }

  .a-inline { container:  inline-size / a; }
  .a-block { container:  block-size / a; }
  .a-size { container:  size / a; }

  .b-inline { container:  inline-size / b; }
  .b-block { container:  block-size / b; }
  .b-size { container:  size / b; }

</style>

<main>
  <div class="inline">
    <div class="block">
      <span></span>
    </div>
  </div>
</main>

<main>
  <div class="block">
    <div class="inline">
      <span></span>
    </div>
  </div>
</main>

<main>
  <div class="size">
    <div class="block">
      <span></span>
    </div>
  </div>
</main>

<main>
  <div class="size">
    <div class="inline">
      <span></span>
    </div>
  </div>
</main>

<main>
  <div class="a-inline">
    <div class="b-inline">
      <span></span>
    </div>
  </div>
</main>

<main>
  <div class="a-inline">
    <div class="a-inline">
      <span></span>
    </div>
  </div>
</main>

<main>
  <div class="a-inline">
    <div class="b-inline">
      <div class="a-block">
        <span></span>
      </div>
    </div>
  </div>
</main>

<script>

  function test_query(prelude, selector, expected) {
    test(t => {
      let elements = document.querySelectorAll(selector);
      assert_equals(elements.length, 1);
      let element = elements[0];

      let style = document.createElement('style');
      t.add_cleanup(() => { style.remove(); });
      style.innerText = `@container ${prelude} { span { --match:true; } } `;
      document.body.append(style);

      assert_equals(getComputedStyle(element).getPropertyValue('--match'), expected);
    }, `${prelude} for ${selector}`);
  }

  // Test that a given container query applies to the specified element.
  // The provided selector must unique identify the element.
  function test_applied(prelude, selector) {
    test_query(prelude, selector, 'true');
  }

  function test_rejected(prelude, selector) {
    test_query(prelude, selector, '');
  }

  // For the following tests, the inner container has a size of 16x16px,
  // and the outer container has a size of 32x32px.

  // "Nearest" selection:
  test_applied('(width: 16px)', '.block > .inline > span');
  test_applied('(height: 16px)', '.inline > .block > span');
  test_rejected('(height)', '.block > .inline > span');
  test_rejected('(width)', '.inline > .block > span');

  // type():
  test_applied('type(inline-size) (width: 32px)', '.inline > .block > span');
  test_applied('type(inline-size) (width: 16px)', '.block > .inline > span');
  test_applied('type(inline-size) (width: 32px)', '.size > .block > span');
  test_applied('type(inline-size) (width: 16px)', '.size > .inline > span');

  test_applied('type(block-size) (height: 16px)', '.inline > .block > span');
  test_applied('type(block-size) (height: 32px)', '.block > .inline > span');
  test_applied('type(block-size) (height: 16px)', '.size > .block > span');
  test_applied('type(block-size) (height: 32px)', '.size > .inline > span');

  test_rejected('type(size) (height)', '.inline > .block > span');
  test_rejected('type(size) (height)', '.block > .inline > span');
  test_rejected('type(size) (width)', '.inline > .block > span');
  test_rejected('type(size) (width)', '.block > .inline > span');
  test_applied('type(size) (height: 32px)', '.size > .block > span');
  test_applied('type(size) (height: 32px)', '.size > .inline > span');

  // name():
  test_applied('name(a) (width: 32px)', '.a-inline > .b-inline > span');
  test_applied('name(b) (width: 16px)', '.a-inline > .b-inline > span');
  test_rejected('name(c) (width)', '.a-inline > .b-inline > span');
  test_applied('name(a) (width: 16px)', '.a-inline > .a-inline > span');

  // Name as plain ident:
  test_applied('a (width: 32px)', '.a-inline > .b-inline > span');
  test_applied('b (width: 16px)', '.a-inline > .b-inline > span');
  test_rejected('c (width)', '.a-inline > .b-inline > span');
  test_applied('a (width: 16px)', '.a-inline > .a-inline > span');

  // The following tests have three containers:
  //
  //  outer  -> 32x32px
  //  middle -> 16x16px
  //  inner -> 8x8px

  // Mixing name() and type():
  test_applied('name(a) type(inline-size) (width: 32px)', '.a-inline > .b-inline > .a-block > span');
  test_applied('name(a) type(block-size) (height: 8px)', '.a-inline > .b-inline > .a-block > span');
  test_rejected('name(a) type(size) (height)', '.a-inline > .b-inline > .a-block > span');
  test_applied('name(b) type(inline-size) (width: 16px)', '.a-inline > .b-inline > .a-block > span');
  test_rejected('name(b) type(block-size) (height)', '.a-inline > .b-inline > .a-block > span');
  test_rejected('name(b) type(size) (height)', '.a-inline > .b-inline > .a-block > span');

  // type() first, then name():
  test_applied('type(inline-size) name(a) (width: 32px)', '.a-inline > .b-inline > .a-block > span');
  test_applied('type(block-size) name(a) (height: 8px)', '.a-inline > .b-inline > .a-block > span');

</script>
