LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/container-selection-expected.txt


FAIL (width: 16px) for .size > .inline > span assert_equals: expected "true" but got ""
FAIL (height: 16px) for .inline > .size > span assert_equals: expected "true" but got ""
FAIL (width: 16px) for .inline > .size > span assert_equals: expected "true" but got ""
FAIL (height: 32px) for .size > .inline > span assert_equals: expected "true" but got ""
PASS (height: 16px) for .size > .inline > span
FAIL a (width: 32px) for .a-size > .b-size > span assert_equals: expected "true" but got ""
FAIL b (width: 16px) for .a-size > .b-size > span assert_equals: expected "true" but got ""
PASS c (width) for .a-size > .b-size > span
FAIL a (width: 16px) for .a-size > .a-size > span assert_equals: expected "true" but got ""
FAIL a (width: 32px) for .a-size > .a > span assert_equals: expected "true" but got ""
FAIL a (width: 32px) for .ab-size > .size > span assert_equals: expected "true" but got ""
FAIL b (width: 32px) for .ab-size > .size > span assert_equals: expected "true" but got ""
PASS c (width) for .ab-size > .size > span
FAIL a (width: 8px) for .a-size > .b-size > .a-inline > span assert_equals: expected "true" but got ""
FAIL b (width: 16px) for .a-size > .b-size > .a-inline > span assert_equals: expected "true" but got ""
FAIL a (height: 32px) for .a-size > .b-size > .a-inline > span assert_equals: expected "true" but got ""
PASS a (height) for .a-inline > .b-size
FAIL a (inline-size: 8px) for .a-size > .b-size > .a-inline > span assert_equals: expected "true" but got ""
FAIL b (inline-size: 16px) for .a-size > .b-size > .a-inline > span assert_equals: expected "true" but got ""
FAIL a (block-size: 32px) for .a-size > .b-size > .a-inline > span assert_equals: expected "true" but got ""
PASS a (block-size) for .a-inline > .b-size
